Move TransferAmountInput to separate file
